Move game code to src
