+++ src/RocketMan.cpp
@@ -4,15 +4,15 @@
 #include <pspthreadman.h>
 #include <cstdlib>
 
-#include "game/game_globals.h"
-#include "game/gamestate.h"
-#include "game/game_utils.h"
+#include "game_globals.h"
+#include "gamestate.h"
+#include "game_utils.h"
 
 #include "globals.h"
 #include "logger.h"
 #include "psp_malloc.h"
 #include "gfx.hpp"
-#include "game/graphics/game_graphics.h"
+#include "graphics/game_graphics.h"
 #include "GUI/text_builder.h"
 
 #include "../Dependencies/FastNoise/FastNoise.hpp"
