+++ src/game_globals.h
