+++ src/game_utils.h
