+++ src/interface/death_menu.h
