+++ src/object_handler/explosion_handler.h
