+++ src/object_handler/global_object_manager.h
